[NPUW] Split MoE builders into model_builder_moe with shared scaffolding
- Factor the Tile-batched MoE scaffolding (input flatten, router-score
  broadcast, expert tiling, weighted expert sum) shared by the GPT-OSS
  and Qwen3 topologies into a BatchedMoEFFN base; new topologies derive
  from it instead of copying the boilerplate.
- Move the MoE FFN builders out of model_builder_ffn into a dedicated
  model_builder_moe, mirroring the per-concern file split of the builder.
- Correct the reshape-shape comment: DeviceRoutedMoETransform does not
  mutate shape Constants in place; it only collects Reshapes whose shape
  input is a Constant and rewires them to new Constants with the expert
  dim set to K.
- Drop the unused device_routed_moe_transform.hpp include from the test.
- Keep the Qwen3 MoE test model stateful (KV cache) like the real model;
  tests that need static shapes now run StatefulToStateless + reshape,
  the same way production prepares an LLM.
