+++ src/plugins/intel_npu/tests/unit/npuw/llm_test_helpers.hpp
@@ -158,18 +158,15 @@ inline std::shared_ptr<ov::Model> build_moe_llm_test_model() {
 /// Qwen3-style MoE: separate gate/up expert MatMuls (SwiGLU) and a Softmax->TopK router
 /// with ReduceSum->Divide renormalization, matching NPUW's Qwen3Expert + Qwen3Router
 /// patterns (real Qwen3-30B-A3B). Contrast build_moe_llm_test_model, which emits the
-/// GPT-OSS topology (fused gate_up, TopK->Softmax router).
-///
-/// The MoE matchers key on the FFN subgraph only, so this model has no KV cache: that
-/// keeps it stateless (no Assign/ReadValue) and lets the partitioner consume it after a
-/// plain static reshape — no StatefulToStateless dance needed.
+/// GPT-OSS topology (fused gate_up, TopK->Softmax router). Stateful (KV cache) like the
+/// real model; consumers that need static shapes run StatefulToStateless + reshape, the
+/// same way production prepares an LLM.
 inline std::shared_ptr<ov::Model> build_qwen3_moe_llm_test_model() {
     ModelBuilder mb;
     auto cfg = make_test_model_config();
     cfg.num_experts = 8;
     cfg.num_experts_per_tok = 2;
     cfg.moe_factory = make_qwen3_moe_ffn;
-    cfg.use_kv_cache = false;  // stateless; MoE patterns don't touch attention/KV
     return mb.build_llm(cfg);
 }
